feat: implement task addition to DOM
Implemented the addTaskToDOM function in the DOMHandler object, to be
able to add a newly create task object to the DOM.

Also added the helper functions, createElement and createInput, to
streamline the process of creating new HTML elements and specifically
HTML input elements. This makes the code for creating a new element of
any type easier to read as it's not a bunch of lines setting the
attributes of the element in a row, but instead just values passed as
arguments to a function to create an element.