fix: remove references of this
Moved the functions to create and delete task and project elements
outside of the return object, and removed any references of 'this'.
Because this is an IIFE, the object doesn't yet exist at the point where
'this' was referenced, so the program would return an error for the
functions being referenced.