refactor: convert factory function to class
Converted the factory function to create Task objects into a class, so
that we can declare static methods to be used on Task objects. This will
allow us to use these functions on Task objects created from a JSON
parse, which was not previously possible since these objects are not
instances of the Task class.