fix: check for projectList length
In the Project class, when creating the ID, it checked if there is an
existing project list, and created and ID based off of the last item in
the list. However, this becomes a problem when projects have been
deleted resulting in the list being just an empty array, while being a
'truthy' value, an error would then be thrown up since we were trying to
access the last item in the array (the array being empty).

Now check that the length of the array is greater than 0 as well.